refactor: :truck: sposta componenti
sposto i componenti seguendo la logica di atomic design
